Use an AtomicInteger for subtemplateCount to avoid race conditions
Right now, the way subtemplate count is implemented is not thread safe
since the ++ operator in java does not atomically read and write the
incremented integer value. Since the field is static this can cause
race conditions leading to a template redefinition since a given
thread may read the same for the integer twice if another thread
reads the value before and later writes the value afterwards.