Generify ModelAdaptor
As suggested in #233

The only change introduced by this that can be considered breaking is the removal of the superclass MapModelAdaptor from AggregateModelAdaptor. The inheritance was replaced by composition. This had to be done because MapModelAdaptor is now a ModelAdaptor<Map>, while AggregateModelAdaptor needs to be a ModelAdaptor<Aggregate>. Previously, AggregateModelAdaptor could not handle a Map, so I think this "breaking" change should not cause any issues.