Fix backward compatibility issues introduced by #233
Reverts NumberRenderer and StringRenderer to accept Object instead of Number or String respectively.

NumberRenderer always worked with Object before.
Actually it should be called FormatRenderer or similar because it merely uses Formatter, nothing Number-specific.

StringRenderer now uses the pre-#233 behaviour of accepting Object at runtime, but failing if the actual value is not a String due to the forced cast.